Update descriptor infrastructure and add support for persistent objects.
Th descriptors no longer support subclasses. Instead custom logic is configured
by passing the name of the custom method to the descriptor constructor.

Persistent objects are used when setting the value of an applookup or
multipleapplookup to a id and when a record gets saved and sync=True is passed
to Record.save().