Add template methods (to `App`, `Record`, `Control` (and subclasses), `LayoutControl` (and subclasses), `Field` (and subclasses), `File`, `Geo`, `MenuItem` and `Panel`).
Add a `globals` attribute to `File` and `Geo`.

Remove `handler` argument from all methods that are visible to UL4.

This means that the handler attached to globals will always be used, but
a handler can be attached to an app for that (which is used by the test
infrastructure to create templates).