Added attribute `View.focus_control`.
Added method `Control.is_focused()`.

Add field errors to the record if we have an active view and the field
is not in the active view. This avoids problems with invisible errors in the
form.

`Control.in_active_view()` is a method now.

Added method `View.focus_first_control()`.

Added method `App.seq()`.